Decode `zappa invoke` output regardless of `--no-color` option
When a user calls `zappa invoke` or `zappa manage`, always decode the
output's `LogResult` as UTF-8, regardless of whether the use has
specified `--no-color`. If the output cannot be decoded as UTF-8, fall
back to printing the raw bytes as a `b''` literal.

Also disables colorization if the output is not being written to a
terminal.