Add example of dash deployment
There's been an lingering issue with using Dash and Zappa producing 403,
404, and 500 errors, depending upon how the app file and Zappa JSON
settings are written. It's difficult to deploy a Dash app with Zappa --
a major bummer, since the combination of Dash's slickness with Zappa's
serverlessness is a killer peanut-butter-and-chocolate combo.

This commit contains a minimal example of a dash app to help users get
started with deploying dash applications using Zappa.

Closes https://github.com/zappa/Zappa/issues/1082